It's possible for the map to load early
In that case, the api-load event would have already fired, so if we're
only listening for that event, we'll never associate the map properties.

Therefore, we have to check if gMap.map exists first. If it does, we can
associate the map properties immediately. Otherwise, we listen for the
api-load event